support pushing a status/color value to a Stoplight widget
See GitHub Issue #5: https://github.com/sonofabell/leftronic/issues/5.

Untested.

This allows you to pass in the thresholds you have set for the particular Stoplight widget you are updating. But it would be better if you could just set the status directly through the API, instead of converting the status to a number.